Apache Hadoop vs other Database technologies
Most database management systems are not up to scratch for operating at such lofty levels of Big Data exigencies either due to the sheer technical inefficiency or the insurmountable financial challenges posed. When the type of data is totally unstructured, the volume of data is humongous, the results needed are at breakneck speeds, then the only platform that can effectively stand up to the challenge is Apache Hadoop.